SW: Add tailspritefree to savegames
git-svn-id: https://svn.eduke32.com/eduke32@8481 1a8010ca-5511-0410-912e-c29ae57300e0

+++ source/sw/src/sounds.cpp
@@ -127,7 +127,7 @@ typedef enum
 char *SongPtr = NULL;
 int SongLength = 0;
 char *SongName = NULL;
-int SongTrack = 0;
+int8_t SongTrack = 0;
 SongType_t SongType = SongTypeNone;
 int SongVoice = -1;
 extern SWBOOL DemoMode;
